Made sure eberything complied with checkstyle. Created a Makefile for easibility, along with a fully documented README.md file. Everything should be all set!

+++ send_signal.c
@@ -1,22 +1,26 @@
 /**
- * @file send_signal.c
- * @brief Program that sends SIGUSR1 along with a random integer value to a target PID.
+ * @file        : send_signal.c
+ * @brief       : Sends SIGUSR1 along with a random integer value to a target PID.
  *
- * Modified by: Jesse Rost <rostj@msoe.edu>
- *
- * Brief summary of program:
+ * Details      :
  * - Takes the target process PID as a command-line argument.
- * - Generates a random integer using rand(), seeded with current time.
- * - Prints the random value before sending.
- * - Sends SIGUSR1 with the integer payload to the receiver using sigqueue().
+ * - Seeds PRNG with current time and generates a random int (0–99).
+ * - Sends SIGUSR1 with integer payload via sigqueue().
  *
- * Compile with:
- *     gcc -Wall -Wextra send_signal.c -o send_signal
+ * Course       : CPE 2600 – Systems Programming
+ * Section      : 112
+ * Assignment   : Signals Lab
+ * Modified by  : Jesse Rost <rostj@msoe.edu>
+ * Date         : 11/10/25
  *
- * Usage:
- *     ./send_signal <pid>
- * Example:
- *     ./send_signal 12345
+ * Compile      : gcc -Wall -Wextra send_signal.c -o send_signal
+ * Usage        : ./send_signal <pid>
+ * Example      : ./send_signal 12345
+ *
+ * Algorithm
+ *  - Parse PID from argv.
+ *  - Seed PRNG and generate a random int.
+ *  - Populate union sigval and call sigqueue().
  */
 
 #define _POSIX_C_SOURCE 200809L
@@ -33,21 +37,21 @@ int main(int argc, char *argv[])
         return 1;
     }
 
-    // Convert command-line argument to PID
+    /* Convert command-line argument to PID */
     pid_t target_pid = (pid_t)atoi(argv[1]);
 
-    // Seed random number generator
-    srand(time(NULL));
+    /* Seed random number generator */
+    srand((unsigned int)time(NULL));
 
-    // Generate random integer
-    int random_value = rand() % 100; // 0–99 range for simplicity
+    /* Generate random integer */
+    int random_value = rand() % 100; /* 0–99 range for simplicity */
     printf("Sending SIGUSR1 with value: %d to PID %d\n", random_value, target_pid);
 
-    // Prepare the union for sigqueue
+    /* Prepare the union for sigqueue */
     union sigval sig_data;
     sig_data.sival_int = random_value;
 
-    // Send the signal
+    /* Send the signal */
     if (sigqueue(target_pid, SIGUSR1, sig_data) == -1)
     {
         perror("sigqueue");
